third_party/blink/web_tests/external/wpt/content-security-policy/securitypolicyviolation/blockeduri-inline-expected.txt

This is a testharness.js-based test.
[FAIL] Inline violations have a blockedURI of 'inline'
  assert_equals: expected 9 but got 0
Harness: the test ran to completion.
